+++ src/test/java/roomescape/controller/api/ReservationApiControllerTest.java
@@ -6,8 +6,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
+import roomescape.domain.ReservationStatus;
+import roomescape.service.dto.request.ReservationSaveRequest;
 import roomescape.util.TokenGenerator;
 
 import java.lang.reflect.Field;
@@ -51,7 +54,8 @@ public class ReservationApiControllerTest {
         Map<String, Object> reservation = Map.of("name", "브라운",
                 "date", LocalDate.now().plusDays(2L).toString(),
                 "timeId", 1,
-                "themeId", 1
+                "themeId", 1,
+                "reservationStatus", "RESERVED"
         );
 
         RestAssured.given().log().all()
@@ -98,5 +102,29 @@ public class ReservationApiControllerTest {
         }
 
         assertThat(isJdbcTemplateInjected).isFalse();
+    }
+
+    @Test
+    @DisplayName("예약 대기 요청을 정상적으로 수행한다.")
+    void createWaiting_Success() {
+        RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(new ReservationSaveRequest(LocalDate.now().plusDays(1L), 1L, 2L, ReservationStatus.WAITING))
+                .cookie("token", TokenGenerator.makeAdminToken())
+                .when().post("/api/reservations")
+                .then().log().all()
+                .statusCode(201);
+    }
+
+    @Test
+    @DisplayName("이미 사용자 본인이 예약 대기 요청을 한 경우 예외를 반환한다.")
+    void alreadyWaitedByUser() {
+        RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(new ReservationSaveRequest(LocalDate.now().plusDays(1L), 1L, 1L, ReservationStatus.WAITING))
+                .cookie("token", TokenGenerator.makeAdminToken())
+                .when().post("/api/reservations")
+                .then().log().all()
+                .statusCode(400);
     }
 }
