+++ src/test/java/roomescape/service/dto/ReservationSaveRequestTest.java
@@ -2,6 +2,7 @@ package roomescape.service.dto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.ReservationStatus;
 import roomescape.service.dto.request.ReservationSaveRequest;
 
 import java.time.LocalDate;
@@ -13,7 +14,7 @@ class ReservationSaveRequestTest {
     @Test
     @DisplayName("이름이 정상 입력될 경우 성공한다.")
     void checkNameBlank_Success() {
-        assertThatCode(() -> new ReservationSaveRequest(LocalDate.now(), 1L, 1L))
+        assertThatCode(() -> new ReservationSaveRequest(LocalDate.now(), 1L, 1L, ReservationStatus.RESERVED))
                 .doesNotThrowAnyException();
     }
 }
